third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/script-charset-01-expected.txt

This is a testharness.js-based test.
PASS Script @type: unknown parameters 
FAIL Script @type: unknown parameters 1 assert_equals: expected "śćążź" but got "ÂśĂ¦ÂąÂżÂź"
PASS Script @type: unknown parameters 2 
FAIL Script @type: unknown parameters 3 assert_equals: expected "śćążź" but got "æ¹¿"
PASS Script @type: unknown parameters 4 
PASS Script @type: unknown parameters 5 
Harness: the test ran to completion.
